build out favCard.tsx
